add pages

+++ src/main/java/com/theBeautiful/model/Price.java
@@ -4,25 +4,25 @@ package com.theBeautiful.model;
  * Created by jiaoli on 10/8/17
  */
 public class Price {
-    private float price;
-    private float salePrice;
+    private Double price;
+    private Double salePrice;
 
     public Price() {
     }
 
-    public float getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
 
-    public float getSalePrice() {
+    public Double getSalePrice() {
         return salePrice;
     }
 
-    public void setSalePrice(float salePrice) {
+    public void setSalePrice(Double salePrice) {
         this.salePrice = salePrice;
     }
 }
